# bank-app/account/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Servicio de Cuentas y Movimientos
  version: 1.0.0
  description: API para gestión de cuentas y movimientos bancarios

servers:
  - url: http://localhost:8080

paths:

  # =====================
  # CUENTAS
  # =====================

  /cuentas:
    get:
      summary: Listar cuentas
      operationId: getCuentas
      responses:
        '200':
          description: Lista de cuentas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cuenta'

    post:
      summary: Crear cuenta
      operationId: createCuenta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuentaCreate'
      responses:
        '201':
          description: Cuenta creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cuenta'
        '400':
          $ref: '#/components/responses/BadRequest'

  /cuentas/{id}:
    get:
      summary: Obtener cuenta por id
      operationId: getCuenta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Cuenta encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cuenta'

    put:
      summary: Actualizar cuenta
      operationId: updateCuenta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            format: int64
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuentaCreate'
      responses:
        '200':
          description: Cuenta actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cuenta'

    patch:
      summary: Actualizar estado cuenta
      operationId: updatePartialCuenta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuentaPartialUpdate'
      responses:
        '200':
          description: Cuenta actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cuenta'

    delete:
      summary: Eliminar cuenta
      operationId: deleteCuenta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Eliminado


  # =====================
  # MOVIMIENTOS
  # =====================

  /movimientos:
    get:
      summary: Listar movimientos
      operationId: getMovimientos
      responses:
        '200':
          description: Lista de movimientos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movimiento'

    post:
      summary: Crear movimiento
      operationId: createMovimiento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovimientoCreate'
      responses:
        '201':
          description: Movimiento creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movimiento'


  /movimientos/{id}:
    get:
      summary: Obtener movimiento
      operationId: getMovimiento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Movimiento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movimiento'


  # =====================
  # REPORTE
  # =====================

  /movimientos/clientes/{clienteId}/reporte:
    get:
      summary: Reporte de estado de cuenta
      operationId: getReporteEstadoCuenta
      parameters:
        - name: clienteId
          in: path
          required: true
          schema:
            type: integer
            format: int64

        - name: fechaInicio
          in: query
          required: true
          schema:
            type: string
            format: date-time

        - name: fechaFin
          in: query
          required: true
          schema:
            type: string
            format: date-time

      responses:
        '200':
          description: Reporte generado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReporteEstadoCuenta'
        '404':
          $ref: '#/components/responses/NotFound'

  /movimientos/clientes/{clienteId}/reporte/pdf:
    get:
      summary: Descargar reporte de estado de cuenta en PDF
      operationId: getReporteEstadoCuentaPdf
      parameters:
        - name: clienteId
          in: path
          required: true
          schema:
            type: integer
            format: int64

        - name: fechaInicio
          in: query
          required: true
          schema:
            type: string
            format: date-time

        - name: fechaFin
          in: query
          required: true
          schema:
            type: string
            format: date-time

      responses:
        '200':
          description: PDF generado
          headers:
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename="reporte.pdf"
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'


components:

  responses:
    BadRequest:
      description: Error de validacion
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:

    # =====================
    # CUENTA
    # =====================

    Cuenta:
      type: object
      properties:
        numeroCuenta:
          type: integer
          format: int64
          readOnly: true
        tipo:
          type: string
        saldoInicial:
          type: number
          format: double
        estado:
          type: boolean
        clienteId:
          type: integer
          format: int64

    CuentaCreate:
      type: object
      required:
        - numero
        - tipo
        - saldoInicial
        - estado
        - clienteId
      properties:

        tipo:
          type: string
          example: "A"

        saldoInicial:
          type: number
          format: double
          example: 200

        estado:
          type: boolean
          example: true

        clienteId:
          type: integer
          format: int64
          example: 1

    CuentaPartialUpdate:
      type: object
      required:
        - estado
      properties:
        estado:
          type: boolean
          example: true


    # =====================
    # MOVIMIENTO
    # =====================

    Movimiento:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fecha:
          type: string
          format: date-time
        tipo:
          type: string
        valor:
          type: number
          format: double
        balance:
          type: number
          format: double
        numeroCuenta:
          type: integer
          format: int64


    MovimientoCreate:
      type: object
      required:
        - fecha
        - tipo
        - valor
        - numeroCuenta
      properties:

        fecha:
          type: string
          format: date-time
          example: "2024-01-01T11:12:55"

        tipo:
          type: string
          example: "deposit"

        valor:
          type: number
          format: double
          example: 200.33

        numeroCuenta:
          type: integer
          format: int64
          example: 1


    # =====================
    # REPORTE
    # =====================

    ReporteEstadoCuenta:
      type: object
      properties:

        fecha:
          type: string
          format: date-time

        cliente:
          type: string

        numeroCuenta:
          type: integer
          format: int64

        tipo:
          type: string

        saldo:
          type: number
          format: double

        estado:
          type: boolean

        movimiento:
          type: array
          items:
            $ref: '#/components/schemas/MovimientoReporte'


    MovimientoReporte:
      type: object
      properties:
        id:
          type: integer

        fecha:
          type: string
          format: date-time

        tipo:
          type: string

        valor:
          type: number
          format: double


    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string